Update javadoc for MessageProducer to change the behaviour when performing an async send when the specified completion listener is null.
Previously this was explicitly allowed and meant the send was async but with no callbacks. 

However this would be inconsistent with the simplified API where setting a null completion listener causes the send to be synchronous. To avoid adding a feature to the old API which would not be available in the simplified API, I have modified the javadoc for MessageProducer to disallow a null completion listener.

svn path=/; revision=127