Small change to spec for MessageProducer when performing an async send. If the completionListener parameter is null an exception should be thrown. Previously it was used to specify that a completion callback was not required.
This change is necessary to avoid a confusing inconsistency with JMSProducer, where calling setAsync(completionListener) with a value of null has the effect of making the subsequent call to send() synchronous. 

Rather than define a null completionListener as having completely different meanings on the two objects I think it is best to define it as having a particular meaning on one object and define it as invalid on the other.

svn path=/; revision=139